rename esnext-browsers.js to browsers.js
